requirements.txt
neo4j==5.20.0
pinecone-client==3.2.2
sentence-transformers==2.7.0
torch>=2.1.0
transformers>=4.35.0
requests==2.32.3
tqdm==4.66.5
python-dotenv==1.0.1
diskcache==5.6.3
pyvis==0.3.2
networkx==3.2.1
steamlit==1.28.0